+++ lib/activerecord/materialized/transaction_refresh_recorder.rb
@@ -9,6 +9,7 @@ module ActiveRecord
       sig { void }
       def initialize
         @tables = T.let([], T::Array[String])
+        @sql_statements = T.let([], T::Array[String])
         @callbacks_registered = T.let(false, T::Boolean)
         @connection = T.let(nil, T.nilable(Connection))
       end
@@ -18,9 +19,10 @@ module ActiveRecord
         @connection = connection
       end
 
-      sig { params(tables: T::Array[String]).void }
-      def add_tables(tables)
+      sig { params(tables: T::Array[String], sql: T.nilable(String)).void }
+      def add_tables(tables, sql: nil)
         @tables.concat(tables)
+        @sql_statements << sql if sql
         register_callbacks! unless @callbacks_registered
       end
 
@@ -33,13 +35,14 @@ module ActiveRecord
         return unless transaction
 
         transaction.after_commit do
-          DependencyRegistry.schedule_refresh_for_tables!(@tables) if @tables.any?
+          DependencyRegistry.schedule_refresh_for_tables!(@tables, sql_statements: @sql_statements) if @tables.any?
         ensure
           DependencyRegistry.clear_recorder(connection)
         end
 
         transaction.after_rollback do
           @tables.clear
+          @sql_statements.clear
         ensure
           DependencyRegistry.clear_recorder(connection)
         end